[SPARK-38036][SQL][TESTS] Refactor `VersionsSuite` to `HiveClientSuite` and make it a subclass of `HiveVersionSuite`
### What changes were proposed in this pull request?
There is a TODO in `VersionsSuite`:

- TODO: Refactor this to `HiveClientSuite` and make it a subclass of `HiveVersionSuite`

this pr completed this TODO, the main change as follows:

- copy all test cases in `versions.foreach` scope of `VersionsSuite` to `HiveClientSuite`
- override `nestedSuites` function in `HiveClientSuites` to use each hive version to test the cases in `HiveClientSuite` similar as `HiveClientUserNameSuites` and `HivePartitionFilteringSuites`
- move other cases to `HiveClientSuites`

### Why are the changes needed?
Make `VersionsSuite` as a subclass of `HiveVersionSuite`  to unify the test mode of multi version hive

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
- Pass GA
- Manual test:

**Before**

```
mvn clean install -DskipTests -pl sql/hive -am
mvn test -pl sql/hive -Dtest=none -DwildcardSuites=org.apache.spark.sql.hive.client.HiveClientSuites

Run completed in 13 minutes, 10 seconds.
Total number of tests run: 867
Suites: completed 2, aborted 0
Tests: succeeded 867, failed 0, canceled 0, ignored 1, pending 0
All tests passed.
```

**After**

```
mvn clean install -DskipTests -pl sql/hive -am
mvn test -pl sql/hive -Dtest=none -DwildcardSuites=org.apache.spark.sql.hive.client.HiveClientSuites

Run completed in 3 minutes, 8 seconds.
Total number of tests run: 867
Suites: completed 14, aborted 0
Tests: succeeded 867, failed 0, canceled 0, ignored 1, pending 0
All tests passed
```

The number of test cases is the same, and Suites changed from 2 to 14

Closes #35335 from LuciferYang/SPARK-38036.

Authored-by: yangjie01 <yangjie01@baidu.com>
